setup mechanism for parsing and matching TEQ subgraphs

+++ experimental/query/search/test/test_trie.cpp
@@ -11,7 +11,7 @@
 #include "experimental/query/search/trie.hpp"
 
 
-const size_t nalphabet = 'z' - 'a' + 1;
+const int nalphabet = 'z' - 'a' + 1;
 
 
 struct AlphaHash final
